Move migration funcs to util package for reuse
This commit moves and exports these functions from migration
packages to util package for reuse by non-migration programs.
- newByAccountRegistersFromPayloadAccountGrouping()
- checkStorageHealth()
- loadAtreeStorageGroup()

For example, these functions can be used by non-migration util
command to run storage health checks.
