+++ cmd/util/cmd/execution-state-extract/execution_state_extract.go
@@ -1,11 +1,9 @@
 package extract
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
 	"os"
-	syncAtomic "sync/atomic"
 	"time"
 
 	"github.com/rs/zerolog"
@@ -15,7 +13,6 @@ import (
 	migrators "github.com/onflow/flow-go/cmd/util/ledger/migrations"
 	"github.com/onflow/flow-go/cmd/util/ledger/reporters"
 	"github.com/onflow/flow-go/cmd/util/ledger/util"
-	"github.com/onflow/flow-go/cmd/util/ledger/util/registers"
 	"github.com/onflow/flow-go/ledger"
 	"github.com/onflow/flow-go/ledger/common/hash"
 	"github.com/onflow/flow-go/ledger/common/pathfinder"
@@ -268,7 +265,7 @@ func newMigration(
 			payloadCount,
 		)
 
-		registersByAccount, err := newByAccountRegistersFromPayloadAccountGrouping(payloadAccountGrouping, nWorker)
+		registersByAccount, err := util.NewByAccountRegistersFromPayloadAccountGrouping(payloadAccountGrouping, nWorker)
 		if err != nil {
 			return nil, err
 		}
@@ -333,99 +330,6 @@ func newMigration(
 
 		return newPayloads, nil
 	}
-}
-
-func newByAccountRegistersFromPayloadAccountGrouping(
-	payloadAccountGrouping *util.PayloadAccountGrouping,
-	nWorker int,
-) (
-	*registers.ByAccount,
-	error,
-) {
-	g, ctx := errgroup.WithContext(context.Background())
-
-	jobs := make(chan *util.PayloadAccountGroup, nWorker)
-	results := make(chan *registers.AccountRegisters, nWorker)
-
-	g.Go(func() error {
-		defer close(jobs)
-		for {
-			payloadAccountGroup, err := payloadAccountGrouping.Next()
-			if err != nil {
-				return fmt.Errorf("failed to group payloads by account: %w", err)
-			}
-
-			if payloadAccountGroup == nil {
-				return nil
-			}
-
-			select {
-			case <-ctx.Done():
-				return ctx.Err()
-			case jobs <- payloadAccountGroup:
-			}
-		}
-	})
-
-	workersLeft := int64(nWorker)
-	for i := 0; i < nWorker; i++ {
-		g.Go(func() error {
-			defer func() {
-				if syncAtomic.AddInt64(&workersLeft, -1) == 0 {
-					close(results)
-				}
-			}()
-
-			for payloadAccountGroup := range jobs {
-
-				// Convert address to owner
-				payloadGroupOwner := flow.AddressToRegisterOwner(payloadAccountGroup.Address)
-
-				accountRegisters, err := registers.NewAccountRegistersFromPayloads(
-					payloadGroupOwner,
-					payloadAccountGroup.Payloads,
-				)
-				if err != nil {
-					return fmt.Errorf("failed to create account registers from payloads: %w", err)
-				}
-				select {
-				case <-ctx.Done():
-					return ctx.Err()
-				case results <- accountRegisters:
-				}
-			}
-
-			return nil
-		})
-	}
-
-	registersByAccount := registers.NewByAccount()
-	g.Go(func() error {
-		for accountRegisters := range results {
-			oldAccountRegisters := registersByAccount.SetAccountRegisters(accountRegisters)
-			if oldAccountRegisters != nil {
-				// Account grouping should never create multiple groups for an account.
-				// In case it does anyway, merge the groups together,
-				// by merging the existing registers into the new ones.
-
-				log.Warn().Msgf(
-					"account registers already exist for account %x. merging %d existing registers into %d new",
-					accountRegisters.Owner(),
-					oldAccountRegisters.Count(),
-					accountRegisters.Count(),
-				)
-
-				err := accountRegisters.Merge(oldAccountRegisters)
-				if err != nil {
-					return fmt.Errorf("failed to merge account registers: %w", err)
-				}
-			}
-		}
-
-		return nil
-	})
-
-	return registersByAccount, g.Wait()
 }
 
 func createTrieFromPayloads(logger zerolog.Logger, payloads []*ledger.Payload) (*trie.MTrie, error) {
